+++ Elena_Poleschuk/Module 3 -  Java/Week 3/Hometask4/src/by/epam/tat/lecture2/task1/objects/SugarCandy.java
