+++ Elena_Poleschuk/Module 3 -  Java/Week 3/Hometask4/src/by/epam/tat/lecture2/task1/utils/ISaver.java
@@ -1,0 +1,12 @@
+package by.epam.tat.lecture2.task1.utils;
+
+import java.util.List;
+
+import by.epam.tat.lecture2.task1.objects.Gift;
+import by.epam.tat.lecture2.task1.objects.Sweets;
+
+public interface ISaver {
+	void saveGift(Gift gitf);
+	List<Sweets> getSavedGift();
+	
+}
