+++ Elena_Poleschuk/Module 3 -  Java/Week 3/Hometask4/src/by/epam/tat/lecture2/task1/utils/exceptions/OpeningSavedCollectionException.java
@@ -1,0 +1,10 @@
+package by.epam.tat.lecture2.task1.utils.exceptions;
+
+@SuppressWarnings("serial")
+public class OpeningSavedCollectionException extends RuntimeException{
+
+	public String getMessage() {
+		return "Gift is not oppened";
+	}	
+	
+}
