--- Elena_Poleschuk/Module 3 -  Java/Week 3/Hometask5/src/by/epam/tat/lecture2/task1/utils/Exceptions.java
@@ -1,17 +1,0 @@
-package by.epam.tat.lecture2.task1.utils;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-
-public class Exceptions {
-	
-	// throw exception on trying to open not existed file
-	public static void isExistFile(String fileName) throws FileNotFoundException {
-	    File file = new File(fileName);
-	    if (!file.exists()){
-	        throw new FileNotFoundException("\"" + file.getName() + "\""+ " file does not exist");
-	    }
-	}
-	
-	
-}
