--- Elena_Poleschuk/Module 3 -  Java/Week 3/Hometask5/src/by/epam/tat/lecture2/task1/utils/IOMethods.java
@@ -1,7 +1,0 @@
-package by.epam.tat.lecture2.task1.utils;
-
-public interface IOMethods {
-	public void saveInfo(String file, String[] collection);
-	public String[][] getSavedGiftInfo(String file);
-	
-}
